requirements.txt
# Core web framework
streamlit>=1.28.0

# Document processing
python-docx>=1.1.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
PyMuPDF>=1.23.0

# Document generation
reportlab>=4.0.0

# LLM integration
openai>=1.3.0
python-dotenv>=1.0.0

# Text processing and analysis
nltk>=3.8.0
scikit-learn>=1.3.0

# Utilities
requests>=2.31.0
Pillow>=10.0.0
